Fix loop counter types from int to size_t for 64-bit
Changed loop counters throughout codebase to use size_t instead of int
when comparing against .size() or container sizes. This prevents potential
signed/unsigned comparison issues on 64-bit systems where size_t is 64-bit
and int is 32-bit.

Files affected:
- Linux: CFormattedTextDisplay.cp (2 instances)
- MacOS: CFormattedTextDisplay.cp, CHTMLTransformer.cp, CPreferencesDialog.cp,
  CCopyToMenu.cp, CAddressFieldSubContainer.cp (9 instances total)
- Win32: CDisplayFormatter.cp, CHTMLTransformer.cp, CServerView.cp, CSMTPView.cp,
  CMailboxView.cp, CMailboxInfoView.cp, CAddressBookView.cp, CAdbkManagerView.cp,
  CAdbkSearchWindow.cp, CCalendarStoreView.cp, CPreferencesDialog.cp,
  CCopyToMenu.cp, CWndAligner.cp, CSDIFrame.cp (18 instances total)
- Sources_Common: CMboxProtocol.h (3 instances - changed offset() parameter too)
- Plug-ins: CKerberosPluginDLL.cp, COutlookExAdbkIOPluginDLL.cp (2 instances)
- Libraries/vCard: CVCardAddress.cpp (2 instances)

Total: 34 instances fixed across 23 files + vCard submodule.

+++ Sources_Common/Mail/Protocols/CMboxProtocol.h
@@ -114,8 +114,8 @@ public:
 		cdstrpairvect mItems[3];
 		size_t size() const
 			{ return mItems[0].size() + mItems[1].size() + mItems[2].size(); }
-		size_t offset(int index)
-			{ int j = 0; for(int i = 0; i < index; i++) j += mItems[i].size(); return j; }
+		size_t offset(size_t index)
+			{ size_t j = 0; for(size_t i = 0; i < index; i++) j += mItems[i].size(); return j; }
 	};
 
 					CMboxProtocol(CMailAccount* account);
